+++ packages/babel-plugin-jsx-sfc/src/utils/ast.ts
@@ -1,21 +1,36 @@
 import * as types from '@babel/types';
-import { JSXElement, JSXIdentifier, JSXAttribute, JSXExpressionContainer, Node } from '@babel/types';
+import { CallExpression } from '@babel/types';
 import { NodePath } from '@babel/traverse';
-
-const LIB_NAME = 'jsx-sfc';
+import { SFC_LIB, SFC_FUNC } from './index';
 
 export function isImportedByLib(
   identifier: string,
-  path: NodePath<JSXElement>,
-  libName: string | string[] = [LIB_NAME]
+  path: NodePath<CallExpression>,
+  libName: string | string[] = [SFC_LIB]
 ) {
   if (!Array.isArray(libName)) {
     libName = [libName];
-    libName.push(LIB_NAME);
+    libName.push(SFC_LIB);
   }
 
   const bindingPath = path.scope.getBinding(identifier)?.path;
   if (types.isImportDeclaration(bindingPath?.parent)) {
     return libName.includes(bindingPath.parent.source.value);
+  }
+}
+
+export function isCalleeImportedBySfc(
+  callee: CallExpression['callee'],
+  path: NodePath<CallExpression>,
+  libName?: string | string[]
+) {
+  if (types.isIdentifier(callee)) {
+    return callee.name === SFC_FUNC && isImportedByLib(callee.name, path, libName);
+  } else if (types.isMemberExpression(callee)) {
+    return (
+      types.isIdentifier(callee.object) &&
+      callee.object.name === SFC_FUNC &&
+      isImportedByLib(callee.object.name, path, libName)
+    );
   }
 }