+++ packages/babel-plugin-jsx-sfc/src/utils/index.ts
@@ -1,0 +1,5 @@
+export const SFC_LIB = 'jsx-sfc';
+
+export const SFC_FUNC = 'sfc';
+
+export const SFC_COMPONENT = 'Component';