+++ packages/jsx-sfc/src/defineComponent.ts
@@ -6,7 +6,7 @@ type NoRef = 'noRef';
 export type JSXElements = ReactElement<any, any> | null;
 
 export type SFCProps<T = {}, EX = {}> = PropsWithChildren<T> & {
-  template: <D extends Template.Data>(data: D) => D;
+  template: <D extends Template.Data>(data?: D) => D;
 } & EX;
 
 export interface SFCInnerProps {
@@ -24,7 +24,7 @@ type ReturnTypeMap<T extends FuncMap> = {
 export type DefineComponent<T = NoRef, P = {}> = {
   <
     D extends Template.Data,
-    S extends Record<string, any>,
+    S,
     EX extends FuncMap,
     SE extends { styles?: S } & ReturnTypeMap<EX>,
     RP = P extends {} ? P : {}
@@ -34,25 +34,20 @@ export type DefineComponent<T = NoRef, P = {}> = {
       Component: T extends NoRef
         ? (props: SFCProps<P, SE>, context?: any) => D
         : (props: SFCProps<P, SE>, ref?: Ref<T>) => D;
-      template: <U extends D>(args: { data: U } & SE) => JSXElements;
+      template: <U extends D>(args: { data: U } & SE, ...tmpls: Template.Func[]) => JSXElements;
     },
     extensions?: EX
   ): (T extends NoRef ? React.FC<RP> : React.ForwardRefExoticComponent<RP & RefAttributes<T>>) & SE;
 
-  <D, S, SP = { styles?: S }, RP = P extends {} ? P : {}>(options: {
-    style?: () => S;
-    Component: T extends NoRef
-      ? (props: SFCProps<P, SP>, context?: any) => D
-      : (props: SFCProps<P, SP>, ref?: Ref<T>) => D;
-    templates: <U extends D>(args: { data: U } & SP, ...tmpls: Template.Func[]) => JSX.Element;
-  }): T extends NoRef ? React.FC<RP> : React.ForwardRefExoticComponent<RP & RefAttributes<T>>;
-
-  <S, SP = { styles?: S }, RP = P extends {} ? P : {}>(options: {
-    style?: () => S;
-    Component: T extends NoRef
-      ? (props: SFCProps<P, SP>, context?: any) => JSXElements
-      : (props: SFCProps<P, SP>, ref?: Ref<T>) => JSXElements;
-  }): T extends NoRef ? React.FC<RP> : React.ForwardRefExoticComponent<RP & RefAttributes<T>>;
+  <S, EX extends FuncMap, SE extends { styles?: S } & ReturnTypeMap<EX>, RP = P extends {} ? P : {}>(
+    options: {
+      style?: () => S;
+      Component: T extends NoRef
+        ? (props: SFCProps<P, SE>, context?: any) => JSXElements
+        : (props: SFCProps<P, SE>, ref?: Ref<T>) => JSXElements;
+    },
+    extensions?: EX
+  ): T extends NoRef ? React.FC<RP> : React.ForwardRefExoticComponent<RP & RefAttributes<T>>;
 };
 
 export interface ForwardRefSFC extends DefineComponent {